Move data classes from logic to model module
