+++ core/database/src/main/sqldelight/com/rendox/routinetracker/core/database/schedule/weekDayMonthRelatedEntity.sq
@@ -1,4 +1,4 @@
-import com.rendox.routinetracker.core.logic.time.WeekDayNumberMonthRelated;
+import com.rendox.routinetracker.core.model.WeekDayNumberMonthRelated;
 import kotlin.Int;
 
 CREATE TABLE weekDayMonthRelatedEntity (
