+++ core/model/build.gradle.kts
@@ -4,7 +4,5 @@ plugins {
 }
 
 dependencies {
-    implementation(project(":core:logic"))
-
     implementation(libs.jetbrains.kotlinx.datetime)
 }
